Add `ChainWorkerState::process_cross_chain_update`
Refactor to move the code from the `WorkerState` into the
`ChainWorkerState`.
