+++ linera-core/src/chain_worker/mod.rs
@@ -6,4 +6,6 @@
 mod config;
 mod state;
 
+#[cfg(test)]
+pub(crate) use self::state::CrossChainUpdateHelper;
 pub use self::{config::ChainWorkerConfig, state::ChainWorkerState};
